Add support for building third party libs as objects. Add more future stb libraries we will use. ImageFormat renamed to PixelFormat with C++11 usage. TextureAddressMode to SamplerAddressMode with C++11 usage. Renderer integration into Engine. Integrate spdlog into Log class which comes with fmt library. Improve README.md details about ThirdParty usage.

+++ Source/Alimer/Renderer/GeometryNode.h
@@ -108,11 +108,11 @@ namespace Alimer
 		/// Set number of geometries.
 		void SetNumGeometries(size_t num);
 		/// Set geometry at index.
-		void SetGeometry(size_t index, Geometry* geometry);
+		void SetGeometry(uint32_t index, Geometry* geometry);
 		/// Set material at every geometry index. Specifying null will use the default material (opaque white.)
 		void SetMaterial(Material* material);
 		/// Set material at geometry index.
-		void SetMaterial(size_t index, Material* material);
+		void SetMaterial(uint32_t index, Material* material);
 		/// Set local space bounding box.
 		void SetLocalBoundingBox(const BoundingBox& box);
